src/requirements.txt
Flask
Flask-RESTX
Flask-Cors
ultralytics
huggingface-hub
easyocr
Pillow
numpy
scikit-learn
python-dotenv
mongoengine
pyjwt
authlib
requests
google-generativeai
opencv-python-headless
flask-socketio
google-genai